Game: cada comida vale el nivel de dificultad actual (_score += _difficulty en step(), no +1 fijo); al poder cambiar la dificultad en caliente, vale la del momento de comer. El puntaje y el record pasan a uint16_t (Game::score/bestScore, Menu::setBestScore y Engine::setBestScore) porque con dificultad 25 y ~96 comidas el maximo teorico (~2400) desbordaba uint8_t (max 255). PROYECTO.md actualizado (secciones 1, 7, 11 y 16)

+++ Engine.cpp
@@ -124,7 +124,7 @@ void Engine::print() {
 // Puntaje máximo (lo conserva el menú entre sesiones)
 // ========================================================
 
-void Engine::setBestScore(uint8_t value) {
+void Engine::setBestScore(uint16_t value) {
   _menu.setBestScore(value);
 }
